# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIRSI AlgoTrader Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BotPing {
  online: boolean;
  /** @nullable */
  status?: string | null;
}

export interface OpenTrade {
  trade_id: number;
  pair: string;
  open_rate: number;
  current_rate: number;
  profit_pct: number;
  profit_abs: number;
  open_date: string;
  stake_amount?: number;
  amount?: number;
  /** @nullable */
  stop_loss_pct?: number | null;
  /** @nullable */
  stoploss_current_dist_pct?: number | null;
}

export interface BotStatus {
  online: boolean;
  /** @nullable */
  mode?: string | null;
  trades?: OpenTrade[];
}

export interface BotProfit {
  /** @nullable */
  profit_all_coin?: number | null;
  /** @nullable */
  profit_all_percent?: number | null;
  /** @nullable */
  profit_closed_coin?: number | null;
  /** @nullable */
  profit_closed_percent?: number | null;
  /** @nullable */
  winning_trades?: number | null;
  /** @nullable */
  losing_trades?: number | null;
  /** @nullable */
  trade_count?: number | null;
  /** @nullable */
  first_trade_date?: string | null;
  /** @nullable */
  latest_trade_date?: string | null;
  /** @nullable */
  avg_duration?: string | null;
  /** @nullable */
  best_pair?: string | null;
  /** @nullable */
  worst_pair?: string | null;
  /** @nullable */
  stake_currency?: string | null;
}

export interface PairPerformance {
  pair: string;
  profit: number;
  profit_pct: number;
  count: number;
}

export interface ClosedTrade {
  trade_id: number;
  pair: string;
  open_rate: number;
  /** @nullable */
  close_rate?: number | null;
  /** @nullable */
  profit_pct?: number | null;
  /** @nullable */
  profit_abs?: number | null;
  open_date: string;
  /** @nullable */
  close_date?: string | null;
  /** @nullable */
  sell_reason?: string | null;
  stake_amount: number;
}

export interface TradeList {
  trades?: ClosedTrade[];
  trades_count?: number;
  total_trades?: number;
}

export interface BalanceCurrency {
  currency: string;
  free: number;
  balance: number;
  used: number;
  est_stake?: number;
  stake?: string;
}

export interface BotBalance {
  currencies?: BalanceCurrency[];
  /** @nullable */
  total?: number | null;
  /** @nullable */
  stake?: string | null;
  /** @nullable */
  note?: string | null;
}

export interface LogLine {
  timestamp: number;
  level: number;
  message: string;
  funcname?: string;
  lineno?: number;
}

export interface BotLogs {
  logs?: LogLine[];
  log_count?: number;
}

export type MarketIntelligenceDecision = {
  generated_at: string;
  expires_at: string;
  allow_long_entries: boolean;
  risk_level: 'normal' | 'guarded' | 'elevated' | 'high';
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  reason: string;
  source_count: number;
  news_count: number;
  model: string;
  snapshot_hash: string;
  errors: string[];
} | null;

export interface MarketIntelligence {
  available: boolean;
  decision: MarketIntelligenceDecision;
}

export interface ExperimentMetrics {
  expectancy: number;
  max_drawdown: number;
  number_of_trades: number;
}

export type ExperimentProposalChanges = {[key: string]: number};

export type ExperimentProposalEvaluationPlan = { [key: string]: unknown };

export type ExperimentProposalSourceSummary = { [key: string]: unknown };

export interface ExperimentProposal {
  schema_version: number;
  proposal_id: string;
  created_at: string;
  title: string;
  hypothesis: string;
  proposal_type: string;
  target_config: string;
  changes: ExperimentProposalChanges;
  evaluation_plan: ExperimentProposalEvaluationPlan;
  source_summary: ExperimentProposalSourceSummary;
  rationale: string;
  status: string;
  [key: string]: unknown;
 }

export type ExperimentEvaluationVerdict = typeof ExperimentEvaluationVerdict[keyof typeof ExperimentEvaluationVerdict];


export const ExperimentEvaluationVerdict = {
  promising: 'promising',
  not_promising: 'not_promising',
  inconclusive: 'inconclusive',
  not_run: 'not_run',
} as const;

export interface ExperimentEvaluation {
  schema_version: number;
  proposal_id: string;
  evaluated_at: string;
  evaluator_version: string;
  baseline: ExperimentMetrics;
  candidate: ExperimentMetrics;
  delta: ExperimentMetrics;
  verdict: ExperimentEvaluationVerdict;
  notes: string;
  [key: string]: unknown;
 }

export type HumanDecisionDecision = typeof HumanDecisionDecision[keyof typeof HumanDecisionDecision];


export const HumanDecisionDecision = {
  approve: 'approve',
  reject: 'reject',
  'request-more-data': 'request-more-data',
} as const;

export interface HumanDecision {
  schema_version: number;
  proposal_id: string;
  decided_at: string;
  reviewer: string;
  decision: HumanDecisionDecision;
  rationale: string;
  apply_to_experimental: boolean;
  /** @nullable */
  applied_path: string | null;
  [key: string]: unknown;
 }

export type ExperimentSummaryStatus = typeof ExperimentSummaryStatus[keyof typeof ExperimentSummaryStatus];


export const ExperimentSummaryStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  'request-more-data': 'request-more-data',
  evaluated: 'evaluated',
} as const;

export interface ExperimentSummary {
  id: string;
  status: ExperimentSummaryStatus;
  proposal: ExperimentProposal;
  evaluation: ExperimentEvaluation | null;
  decision: HumanDecision | null;
  /** @nullable */
  experimental_profile: string | null;
}

export interface ExperimentList {
  experiments: ExperimentSummary[];
  total: number;
}

export type ExperimentDecisionInputDecision = typeof ExperimentDecisionInputDecision[keyof typeof ExperimentDecisionInputDecision];


export const ExperimentDecisionInputDecision = {
  approve: 'approve',
  reject: 'reject',
  'request-more-data': 'request-more-data',
} as const;

export interface ExperimentDecisionInput {
  decision: ExperimentDecisionInputDecision;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  reviewer?: string | null;
  apply_experimental?: boolean;
}

export interface BotConfig {
  /** @nullable */
  strategy?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  dry_run?: boolean | null;
  /** @nullable */
  stake_currency?: string | null;
  /** @nullable */
  stake_amount?: number | null;
  /** @nullable */
  max_open_trades?: number | null;
  /** @nullable */
  timeframe?: string | null;
  /** @nullable */
  exchange?: string | null;
}

export type BotTradesParams = {
limit?: number;
};

export type BotLogsParams = {
limit?: number;
};

export type ListExperimentsParams = {
/**
 * @minimum 1
 * @maximum 200
 */
limit?: number;
};