# lib/api-zod/src/generated/types/experimentDecisionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIRSI AlgoTrader Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ExperimentDecisionInputDecision } from './experimentDecisionInputDecision';

export interface ExperimentDecisionInput {
  decision: ExperimentDecisionInputDecision;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  reviewer?: string | null;
  apply_experimental?: boolean;
}