# lib/api-zod/src/generated/types/experimentEvaluation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIRSI AlgoTrader Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ExperimentEvaluationVerdict } from './experimentEvaluationVerdict';
import type { ExperimentMetrics } from './experimentMetrics';

export interface ExperimentEvaluation {
  schema_version: number;
  proposal_id: string;
  evaluated_at: string;
  evaluator_version: string;
  baseline: ExperimentMetrics;
  candidate: ExperimentMetrics;
  delta: ExperimentMetrics;
  verdict: ExperimentEvaluationVerdict;
  notes: string;
  [key: string]: unknown;
 }