# lib/api-zod/src/generated/types/experimentList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIRSI AlgoTrader Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ExperimentSummary } from './experimentSummary';

export interface ExperimentList {
  experiments: ExperimentSummary[];
  total: number;
}