# lib/api-zod/src/generated/types/experimentMetrics.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIRSI AlgoTrader Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ExperimentMetrics {
  expectancy: number;
  max_drawdown: number;
  number_of_trades: number;
}