# lib/api-zod/src/generated/types/experimentProposal.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIRSI AlgoTrader Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ExperimentProposalChanges } from './experimentProposalChanges';
import type { ExperimentProposalEvaluationPlan } from './experimentProposalEvaluationPlan';
import type { ExperimentProposalSourceSummary } from './experimentProposalSourceSummary';

export interface ExperimentProposal {
  schema_version: number;
  proposal_id: string;
  created_at: string;
  title: string;
  hypothesis: string;
  proposal_type: string;
  target_config: string;
  changes: ExperimentProposalChanges;
  evaluation_plan: ExperimentProposalEvaluationPlan;
  source_summary: ExperimentProposalSourceSummary;
  rationale: string;
  status: string;
  [key: string]: unknown;
 }