# lib/api-zod/src/generated/types/experimentSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIRSI AlgoTrader Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ExperimentEvaluation } from './experimentEvaluation';
import type { ExperimentProposal } from './experimentProposal';
import type { ExperimentSummaryStatus } from './experimentSummaryStatus';
import type { HumanDecision } from './humanDecision';

export interface ExperimentSummary {
  id: string;
  status: ExperimentSummaryStatus;
  proposal: ExperimentProposal;
  evaluation: ExperimentEvaluation | null;
  decision: HumanDecision | null;
  /** @nullable */
  experimental_profile: string | null;
}