# lib/api-zod/src/generated/types/humanDecision.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIRSI AlgoTrader Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { HumanDecisionDecision } from './humanDecisionDecision';

export interface HumanDecision {
  schema_version: number;
  proposal_id: string;
  decided_at: string;
  reviewer: string;
  decision: HumanDecisionDecision;
  rationale: string;
  apply_to_experimental: boolean;
  /** @nullable */
  applied_path: string | null;
  [key: string]: unknown;
 }